refactor(activity): rename layers to core, queries, components
Frontend-native vocabulary for the layered feature layout. No behavior
change. FE-31 rule globs and regexes follow the new paths.
