+++ apps/web/src/features/activity/core/activity-stats.ts
