+++ apps/web/src/features/activity/core/group-events.test.ts
