+++ apps/web/src/features/activity/core/intensity.test.ts
