+++ apps/web/src/features/activity/core/intensity.ts
