+++ apps/web/src/features/activity/queries/entity-query.ts
