+++ rules/ast-grep/ts-no-activity-domain-framework-imports.yml
@@ -5,17 +5,17 @@ language: typescript
 severity: warning
 message: Activity domain modules must stay pure TypeScript — no Solid, urql, queries, generated GraphQL, or feature adapters.
 note: |-
-  Rule FE-31 in docs/STYLE_GUIDE.md. `features/activity/domain` owns the
+  Rule FE-31 in docs/STYLE_GUIDE.md. `features/activity/core` owns the
   activity vocabulary and the functions that describe, group, and summarize
   it. Wire types, clients, and UI stay outside this directory so a schema
   rename cannot churn domain tests. Decode and query factories live in
-  `features/activity/adapters`. The `dateBucket` import from next-soup is
+  `features/activity/queries`. The `dateBucket` import from next-soup is
   the documented exception for feed labels that must match soup views.
 files:
-  - apps/web/src/features/activity/domain/**
+  - apps/web/src/features/activity/core/**
 rule:
   kind: string
-  regex: ^['"](solid-js|@urql|urql|@app/lib/urql|@queries/|@service-storage/graphql/generated/|(\.\./)+adapters/|@app/features/activity/adapters)
+  regex: ^['"](solid-js|@urql|urql|@app/lib/urql|@queries/|@service-storage/graphql/generated/|(\.\./)+queries/|@app/features/activity/queries)
   any:
     - inside:
         kind: import_statement
